tweaks in rezising & filesystem structure
